requirements.txt
streamlit==0.89.0
python-dotenv==0.19.2
google-generativeai==0.1.3
